requirements.txt
# FastAPI and server
fastapi==0.120.2
uvicorn==0.23.2

# TensorFlow + Keras
keras==2.13.1

# Data processing
numpy==1.24.3
pillow==10.4.0

# For UI / optional
gradio==3.46.1

# Optional: logging, utils
python-multipart==0.0.6
